backend/requirements.txt
anyio==4.13.0
asgiref==3.11.1
channels==4.3.2
channels_redis==4.3.0
click==8.3.1
Django==6.0.3
django-ipware==7.0.1
django-structlog==10.0.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
h11==0.16.0
httptools==0.7.1
idna==3.11
iniconfig==2.3.0
msgpack==1.1.2
packaging==26.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.12.1
pytest==9.0.2
pytest-django==4.12.0
python-dotenv==1.2.2
python-ipware==3.0.0
PyYAML==6.0.3
redis==7.4.0
sqlparse==0.5.5
structlog==25.5.0
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
